dodo-server/requirements.txt
aniso8601==9.0.1
black==22.1.0
blinker==1.4
cassandra-driver==3.25.0
click==8.0.3
cmake==3.22.4
colorama==0.4.4
cycler==0.11.0
distlib==0.3.4
dlib==19.23.1
filelock==3.6.0
Flask==2.2.2
Flask-Cors==3.0.9
Flask-CQLAlchemy==2.0.0
Flask-JSON==0.3.4
Flask-JWT-Extended==4.4.3
Flask-RESTful==0.3.8
flask-restful-swagger-2==0.35
fonttools==4.30.0
geomet==0.2.1.post1
gunicorn==19.10.0
imutils==0.5.4
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.1.1
matplotlib==3.5.1
mypy-extensions==0.4.3
mysql-connector-python==8.0.28
neo4j==4.1.1
numpy==1.22.3
opencv-contrib-python==4.5.5.64
packaging==21.3
passlib==1.7.4
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.1
PyJWT==2.4.0
pymongo==4.2.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.14.0
pytz==2021.3
six==1.16.0
tomli==2.0.1
virtualenv==20.13.3
Werkzeug==2.2.2
